refactor: factorize code
