+++ src/lib/helpers/string.ts
@@ -8,4 +8,14 @@ function vec3ToString(vec3: THREE.Vector3Like, separator: string = 'x'): string 
     return `${vec3.x}${separator}${vec3.y}${separator}${vec3.z}`;
 }
 
-export { vec2ToString, vec3ToString };
+function applyReplacements(source: string, replacements: Record<string, string>): string {
+    let result = source;
+
+    for (const [source, replacement] of Object.entries(replacements)) {
+        result = result.replace(source, replacement);
+    }
+
+    return result;
+}
+
+export { applyReplacements, vec2ToString, vec3ToString };
